lab5

+++ controler.py
@@ -1,9 +1,11 @@
-from flask import json, Response, request
+from flask import json, Response, request, jsonify
 from flask_restful import Resource, Api
 from models import *
 from werkzeug.security import generate_password_hash, check_password_hash
 from sqlalchemy.ext.declarative import DeclarativeMeta
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app import session
+
 class AlchemyEncoder(json.JSONEncoder):
 
     def default(self, obj):
@@ -23,15 +25,20 @@ class AlchemyEncoder(json.JSONEncoder):
         return json.JSONEncoder.default(self, obj)
 
 
-
-
-
 class AddShow(Resource):
+    @jwt_required
     def post(self):
         data = request.json
-
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin != 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for users"}),
+                status=400,
+                mimetype="application/json"
+            )
         try:
-            show = Show(data["name"],data["show_type"],data["description"],data["time"],data["place"])
+            show = Show(data["name"], data["show_type"], data["description"], data["time"], data["place"])
             session.add(show)
             session.flush()
             session.commit()
@@ -49,6 +56,7 @@ class AddShow(Resource):
 
 
 class GetAllShows(Resource):
+    @jwt_required
     def get(self):
         shows = session.query(Show).all()
         if shows:
@@ -64,12 +72,11 @@ class GetAllShows(Resource):
             )
 
 
-
 class SignUpUser(Resource):
     def post(self):
         data = request.json
         try:
-            user = User(data["name"],data["password"],data["phone"],data["mail"])
+            user = User(data["name"], data["password"], data["phone"], data["mail"], data["admin"])
             checkuser = session.query(User).filter(User.mail == user.mail).all()
             if checkuser:
                 return Response(
@@ -94,8 +101,43 @@ class SignUpUser(Resource):
             )
 
 
-class GetAllUsers(Resource):
+class Login(Resource):
+    def post(self):
+        data = request.json
+        user = User.authenticate(**data)
+        token = user.get_token()
+        return jsonify({'access_token': token})
+
+
+class GetOwnUser(Resource):
+    @jwt_required
     def get(self):
+        id = get_jwt_identity()
+        user = session.query(User).get(id)
+        if user:
+            return Response(
+                response=json.dumps(user, cls=AlchemyEncoder),
+                status=201,
+                mimetype="application/json"
+            )
+        return Response(
+            response=json.dumps({"message": "Not found"}),
+            status=404,
+            mimetype="application/json"
+        )
+
+
+class GetAllUsers(Resource):
+    @jwt_required
+    def get(self):
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin != 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for users"}),
+                status=400,
+                mimetype="application/json"
+            )
         users = session.query(User).all()
         if users:
             return Response(
@@ -111,8 +153,55 @@ class GetAllUsers(Resource):
 
 
 class DeleteUser(Resource):
+    @jwt_required
     def delete(self, id):
-        user = session.query(User).filter(User.id==id).delete()
+        try:
+            user_id = get_jwt_identity()
+            checkuser = session.query(User).filter(User.id == user_id).one()
+            if checkuser.admin != 1:
+                return Response(
+                    response=json.dumps({"message": "Not allowed for users"}),
+                    status=400,
+                    mimetype="application/json"
+                )
+            tickets = session.query(Ticket).filter(Ticket.user_id == id).all()
+            for ticket in tickets:
+                ticket.is_avaliable = 1
+                ticket.user_id = None
+            session.commit()
+            user = session.query(User).filter(User.id == id).delete()
+            session.commit()
+            if user:
+                return Response(
+                    response=json.dumps({"message": "Success"}),
+                    status=200,
+                    mimetype="application/json"
+                )
+        except Exception as e:
+            return Response(
+                response=json.dumps({"message": "Not found"}),
+                status=400,
+                mimetype="application/json"
+            )
+
+
+class DeleteMyself(Resource):
+    @jwt_required
+    def delete(self):
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin == 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for admin"}),
+                status=400,
+                mimetype="application/json"
+            )
+        tickets = session.query(Ticket).filter(Ticket.user_id == user_id).all()
+        for ticket in tickets:
+            ticket.is_avaliable = 1
+            ticket.user_id = None
+        session.commit()
+        user = session.query(User).filter(User.id == id).delete()
         session.commit()
         if user:
             return Response(
@@ -127,7 +216,16 @@ class DeleteUser(Resource):
             )
 
 class GetUser(Resource):
+    @jwt_required
     def get(self, id):
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin != 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for users"}),
+                status=400,
+                mimetype="application/json"
+            )
         user = session.query(User).get(id)
         if user:
             return Response(
@@ -142,10 +240,50 @@ class GetUser(Resource):
             )
 
 
+class UpdateUser(Resource):
+    @jwt_required
+    def put(self):
+        data = request.json
+        user_id = get_jwt_identity()
+        try:
+            user = session.query(User).get(user_id)
+            if not user:
+                return Response(
+                    response=json.dumps({"message": "invalid id"}),
+                    status=400,
+                    mimetype="application/json"
+                )
+            if "password" in data:
+                user.password = generate_password_hash(data['password'])
+            if "name" in data:
+                user.name = data["name"]
+            if "phone" in data:
+                user.phone = data["phone"]
+            session.commit()
+            return Response(
+                response=json.dumps({"message": "Success"}),
+                status=200,
+                mimetype="application/json"
+            )
+        except Exception as e:
+            return Response(
+                response=json.dumps({"message": "Invalid input"}),
+                status=405,
+                mimetype="application/json"
+            )
+
 
 class AddTicket(Resource):
+    @jwt_required
     def post(self, sid):
-
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin != 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for users"}),
+                status=400,
+                mimetype="application/json"
+            )
         data = request.json
         try:
             ticket = Ticket(1, data["clas"], sid, None)
@@ -166,9 +304,18 @@ class AddTicket(Resource):
                 mimetype="application/json"
             )
 
-class BuyTicket(Resource):
-    def put(self, tid, uid):
 
+class BuyTicket(Resource):
+    @jwt_required
+    def put(self, tid):
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin == 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for admins"}),
+                status=400,
+                mimetype="application/json"
+            )
         try:
             ticket = session.query(Ticket).get(tid)
             if not ticket.is_avaliable:
@@ -178,7 +325,7 @@ class BuyTicket(Resource):
                     mimetype="application/json"
                 )
             ticket.is_avaliable = 0
-            ticket.user_id = uid
+            ticket.user_id = user_id
 
             session.commit()
             return Response(
@@ -196,8 +343,16 @@ class BuyTicket(Resource):
 
 
 class ReserveTicket(Resource):
-    def put(self, tid, uid):
-
+    @jwt_required
+    def put(self, tid):
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin == 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for admins"}),
+                status=400,
+                mimetype="application/json"
+            )
         try:
             ticket = session.query(Ticket).get(tid)
             if not ticket.is_avaliable:
@@ -207,7 +362,7 @@ class ReserveTicket(Resource):
                     mimetype="application/json"
                 )
             ticket.is_avaliable = 0
-            ticket.user_id = uid
+            ticket.user_id = user_id
 
             session.commit()
             return Response(
@@ -222,11 +377,19 @@ class ReserveTicket(Resource):
                 mimetype="application/json"
             )
 
-class DeleteReservationTicket(Resource):
-    def put(self, tid):
 
+class DeleteReservationTicket(Resource):
+    @jwt_required
+    def put(self, tid):
         try:
+            user_id = get_jwt_identity()
             ticket = session.query(Ticket).get(tid)
+            if ticket.user_id != user_id:
+                return Response(
+                    response=json.dumps({"message": "Not allowed"}),
+                    status=400,
+                    mimetype="application/json"
+                )
             ticket.is_avaliable = 1
             ticket.user_id = None
 
@@ -245,23 +408,48 @@ class DeleteReservationTicket(Resource):
 
 
 class GetTicket(Resource):
+    @jwt_required
     def get(self, id):
-        ticket = session.query(Ticket).get(id)
-        if ticket:
+        try:
+            user_id = get_jwt_identity()
+            checkuser = session.query(User).filter(User.id == user_id).one()
+            if checkuser.admin != 1:
+                return Response(
+                    response=json.dumps({"message": "Not allowed for user"}),
+                    status=400,
+                    mimetype="application/json"
+                )
+            ticket = session.query(Ticket).get(id)
+            if ticket:
+                return Response(
+                    response=json.dumps(ticket, cls=AlchemyEncoder),
+                    status=201,
+                    mimetype="application/json"
+                )
             return Response(
-                response=json.dumps(ticket, cls=AlchemyEncoder),
-                status=201,
+                    response=json.dumps({"message": "Not found"}),
+                    status=404,
+                    mimetype="application/json"
+                )
+        except Exception as e:
+            return Response(
+                response=json.dumps({"message": "Invalid input"}),
+                status=405,
                 mimetype="application/json"
             )
-        return Response(
-                response=json.dumps({"message": "Not found"}),
-                status=404,
-                mimetype="application/json"
-            )
+
 
 class GetAllTickets(Resource):
-
+    @jwt_required
     def get(self):
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin != 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for users"}),
+                status=400,
+                mimetype="application/json"
+            )
         ticket = session.query(Ticket).all()
         if ticket:
             return Response(
@@ -277,7 +465,35 @@ class GetAllTickets(Resource):
 
 
 class GetAllUserTickets(Resource):
+    @jwt_required
     def get(self, uid):
+        user_id = get_jwt_identity()
+        checkuser = session.query(User).filter(User.id == user_id).one()
+        if checkuser.admin != 1:
+            return Response(
+                response=json.dumps({"message": "Not allowed for users"}),
+                status=400,
+                mimetype="application/json"
+            )
+        ticket = session.query(Ticket).filter(Ticket.user_id == uid).all()
+
+        if ticket:
+            return Response(
+                response=json.dumps(ticket, cls=AlchemyEncoder),
+                status=201,
+                mimetype="application/json"
+            )
+        return Response(
+            response=json.dumps({"message": "Not found"}),
+            status=404,
+            mimetype="application/json"
+        )
+
+
+class GetMyTickets(Resource):
+    @jwt_required
+    def get(self):
+        uid = get_jwt_identity()
         ticket = session.query(Ticket).filter(Ticket.user_id == uid).all()
 
         if ticket:
